refactor: split server.js starter to different files
