--- src/app/multi.ts
@@ -1,33 +1,0 @@
-import cluster from "node:cluster";
-import http from "node:http";
-import { cpus } from "node:os";
-import process from "node:process";
-
-import * as dotenv from 'dotenv';
-dotenv.config();
-
-const numCPUs = cpus().length;
-
-if (cluster.isPrimary) {
-  console.log(`Primary ${process.pid} is running`);
-
-  // Fork workers.
-  for (let i = 0; i < numCPUs; i++) {
-    cluster.fork();
-  }
-
-  cluster.on("exit", (worker, code, signal) => {
-    console.log(`worker ${worker.process.pid} died`);
-  });
-} else {
-  // Workers can share any TCP connection
-  // In this case it is an HTTP server
-  http
-    .createServer((req, res) => {
-      res.writeHead(200);
-      res.end("hello world\n");
-    })
-    .listen(process.env.MULTIPORT);
-
-  console.log(`Worker ${process.pid} started`);
-}
